+++ src/controller/FloorDetector.lua
@@ -6,18 +6,23 @@ local Vec3              = require("math/Vec3")
 ---@field Instance fun():FloorDetector
 ---@field Measure fun():TelemeterResult
 ---@field Present fun():boolean
+---@field MaxDist fun():number
+---@field EnableParking fun(on:boolean)
+---@field IsWithinShutdownDistance fun():boolean
+---@field IsParkingEnabled fun():boolean
 
 local FloorDetector     = {}
 FloorDetector.__index   = FloorDetector
 
-local instance
+local inst
 local floorDetectorName = "FloorDetector"
 
 ---@return FloorDetector
 function FloorDetector.Instance()
-    if instance then return instance end
+    if inst then return inst end
 
-    instance = {}
+    inst = {}
+    local enabled = false
 
     local teleLink = library.getLinkByName(floorDetectorName)
     local tele ---@type Telemeter|nil
@@ -31,12 +36,12 @@ function FloorDetector.Instance()
         log.Error("No telementer by name '", floorDetectorName, "' found")
     end
 
-    function instance.Present()
+    function inst.Present()
         return tele ~= nil
     end
 
     ---@return TelemeterResult
-    function instance.Measure()
+    function inst.Measure()
         if tele then
             return tele.Measure()
         else
@@ -44,7 +49,21 @@ function FloorDetector.Instance()
         end
     end
 
-    return setmetatable(instance, FloorDetector)
+    ---@return number #Max distance or 0
+    function inst.MaxDist()
+        return tele and tele.MaxDistance() or 0
+    end
+
+    ---@param v boolean
+    function inst.EnableParking(v)
+        enabled = v
+    end
+
+    function inst.IsParkingEnabled()
+        return enabled
+    end
+
+    return setmetatable(inst, FloorDetector)
 end
 
 return FloorDetector