+++ src/controller/Wsad.lua
@@ -9,6 +9,7 @@ local constants               = s.constants
 local defaultMargin           = constants.flight.defaultMargin
 local calc                    = s.calc
 local keys                    = s.keys
+local floor                   = s.floorDetector
 local VerticalReferenceVector = universe.VerticalReferenceVector
 local Sign                    = s.calc.Sign
 local NF                      = function() return not IsFrozen() end
@@ -235,6 +236,7 @@ function Wsad.New(fsm, flightCore, settings, access)
         end
 
         newMovement = value ~= prev
+        floor.EnableParking(false)
 
         return value
     end